// src/app/models/profile.model.ts
export interface Profile {
  name: string
  profession: string
}

export interface Experience {
  title: string
  period: string
  location: string
  description: string
}

export interface Skill {
  name: string
  level: number
}

export interface Language {
  name: string
}

export interface Hobby {
  name: string
}

export interface Contact {
  phone: string
  website: string
  address: string
}

export interface Education {
  institution: string
  degree: string
  period: string
}

export interface Reference {
  name: string
  address: string
  phone: string
  email: string
}